the 'zom_parser' package has been flatten, and dependencies are resolved TODO: * we could probably break the lib.rs of zom_lexer into multiple files, instead of having one big file ?
