+++ Cargo.toml
@@ -18,3 +18,8 @@ readme = "README.md"
 [workspace.dependencies]
 inkwell = { version = "0.2.0", default-features = false }
 criterion = "0.5.0"
+zom_lexer = { path = "stage1/zom_lexer" }
+zom_parser = { path = "stage1/zom_parser" }
+zom_common = { path = "stage1/zom_common" }
+zom_compiler = { path = "stage1/zom_compiler" }
+zom_codegen = { path = "stage1/zom_codegen" }
