+++ stage1/zom/src/ops/bobj.rs
@@ -8,10 +8,11 @@ use std::{
 use inkwell::{context::Context, passes::PassManager};
 // use zom_codegen::gen::CodeGen;
 // use zom_compiler::compiler::Compiler;
-use zom_fe::{
-    lexer::Lexer,
-    parser::{parse, ParserSettings, ParsingContext},
+use zom_parser::{
+    parse, ParserSettings, ParsingContext
 };
+
+use zom_lexer::Lexer;
 
 use crate::{err, ExitStatus};
