+++ stage1/zom_lexer/Cargo.toml
@@ -11,4 +11,4 @@ authors.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-zom_common = { path = "../zom_common" }
+zom_common.workspace = true
