+++ stage1/zom_parser/src/item.rs
@@ -1,17 +1,17 @@
 //! Module responsible for parsing items, function definition, function declaration or global symbols.
 
 use crate::parse_try;
-use crate::parser::symbol::Symbol;
-use crate::parser::CodeLocation;
-use crate::{err_et, parser::function::Function};
+use crate::symbol::Symbol;
+use crate::CodeLocation;
+use crate::{err_et, function::Function};
 
-use crate::parser::PartParsingResult::*;
+use crate::PartParsingResult::*;
 use zom_common::token::{Token, TokenType::*};
 
 use core::ops::Range;
 
-use super::symbol::parse_global_symbol;
-use super::{
+use crate::symbol::parse_global_symbol;
+use crate::{
     function::{parse_extern, parse_function},
     ParserSettings, ParsingContext, PartParsingResult,
 };
