+++ stage1/zom_parser/src/types.rs
